requirements.txt
openai
Django
python-decouple
djangorestframework